requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2